refactor: [#31] 패키지명 오타 수정 contoller -> controller
