+++ src/main/java/com/toy/namoner/domain/user/service/UserService.java
@@ -2,8 +2,8 @@ package com.toy.namoner.domain.user.service;
 
 import java.util.Optional;
 
-import com.toy.namoner.domain.user.contoller.dto.request.UserInfoUpdateRequest;
-import com.toy.namoner.domain.user.contoller.dto.response.UserInfoUpdateResponse;
+import com.toy.namoner.domain.user.controller.dto.request.UserInfoUpdateRequest;
+import com.toy.namoner.domain.user.controller.dto.response.UserInfoUpdateResponse;
 import com.toy.namoner.domain.user.model.User;
 
 public interface UserService {
